+++ packages/web-components/fast-components/rollup.config.js
@@ -1,3 +1,4 @@
+/* eslint-env node */
 import commonJS from "rollup-plugin-commonjs";
 import filesize from "rollup-plugin-filesize";
 import resolve from "rollup-plugin-node-resolve";
@@ -17,6 +18,15 @@ export default [
     {
         context: "this",
         input: "src/index-rollup.ts",
+        onwarn(warning, warn) {
+            // The IIFE export doesn't have a namespace since component exports
+            // are expected to be top-level objects
+            if (warning.code === "MISSING_NAME_OPTION_FOR_IIFE_EXPORT") {
+                return;
+            }
+
+            warn(warning);
+        },
         output: [
             {
                 file: "dist/fast-components.js",