+++ packages/web-components/fast-components/src/disclosure/disclosure.stories.ts
@@ -1,0 +1,13 @@
+import { FASTDesignSystemProvider } from "../design-system-provider";
+import DisclosureTemplate from "./fixtures/disclosure.html";
+import { FASTDisclosure } from ".";
+
+// Prevent tree-shaking
+FASTDisclosure;
+FASTDesignSystemProvider;
+
+export default {
+    title: "Disclosure",
+};
+
+export const disclosure = () => DisclosureTemplate;