+++ packages/web-components/fast-components/src/text-area/text-area.definition.ts
@@ -1,4 +1,4 @@
-import { WebComponentDefinition } from "@microsoft/fast-tooling/dist/data-utilities/web-component";
+import { WebComponentDefinition } from "@microsoft/fast-tooling/dist/esm/data-utilities/web-component";
 import { DataType } from "@microsoft/fast-tooling";
 import { TextAreaResize } from "@microsoft/fast-foundation/dist/esm/text-area/text-area.options";