+++ packages/web-components/fast-components/src/text-field/text-field.definition.ts
@@ -1,4 +1,4 @@
-import { WebComponentDefinition } from "@microsoft/fast-tooling/dist/data-utilities/web-component";
+import { WebComponentDefinition } from "@microsoft/fast-tooling/dist/esm/data-utilities/web-component";
 import { DataType } from "@microsoft/fast-tooling";
 import { TextFieldType } from "@microsoft/fast-foundation/dist/esm/text-field/text-field.options";