+++ packages/tooling/fast-figma-plugin-msft/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.9](https://github.com/Microsoft/fast/compare/@microsoft/fast-figma-plugin-msft@0.8.8...@microsoft/fast-figma-plugin-msft@0.8.9) (2021-01-30)
+
+**Note:** Version bump only for package @microsoft/fast-figma-plugin-msft
+
+
+
+
+
 ## [0.8.8](https://github.com/Microsoft/fast/compare/@microsoft/fast-figma-plugin-msft@0.8.7...@microsoft/fast-figma-plugin-msft@0.8.8) (2020-12-17)
 
 **Note:** Version bump only for package @microsoft/fast-figma-plugin-msft