+++ packages/tooling/fast-tooling-react/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.3](https://github.com/Microsoft/fast/compare/@microsoft/fast-tooling-react@2.5.2...@microsoft/fast-tooling-react@2.5.3) (2021-01-30)
+
+**Note:** Version bump only for package @microsoft/fast-tooling-react
+
+
+
+
+
 ## [2.5.2](https://github.com/Microsoft/fast/compare/@microsoft/fast-tooling-react@2.5.1...@microsoft/fast-tooling-react@2.5.2) (2020-12-16)
 
 **Note:** Version bump only for package @microsoft/fast-tooling-react