+++ sites/fast-website/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.9](https://github.com/Microsoft/fast/compare/@microsoft/fast-website@1.3.8...@microsoft/fast-website@1.3.9) (2021-01-30)
+
+**Note:** Version bump only for package @microsoft/fast-website
+
+
+
+
+
 ## [1.3.8](https://github.com/Microsoft/fast/compare/@microsoft/fast-website@1.3.7...@microsoft/fast-website@1.3.8) (2020-12-17)
 
 **Note:** Version bump only for package @microsoft/fast-website